+++ spec/apps/kitchen_sink/app/main/models/todo.rb
@@ -1,0 +1,3 @@
+class Todo < Volt::Model
+  field :name, String
+end