+++ spec/server/rack/asset_files_spec.rb
@@ -3,7 +3,7 @@ if RUBY_PLATFORM != 'opal'
   require 'volt/server/rack/asset_files'
   require 'volt/server/template_handlers/sprockets_component_handler'
 
-   describe Volt::AssetFiles do
+  describe Volt::AssetFiles do
     before do
       spec_app_root = File.join(File.dirname(__FILE__), '../../apps/file_loading')
 
@@ -14,38 +14,42 @@ if RUBY_PLATFORM != 'opal'
       main = Volt::AssetFiles.new('/app', 'main', @component_paths)
 
       components = main.components
-      expect(components).to eq(%w(volt postgres main shared bootstrap slideshow))
+      expect(components).to eq(%w(volt sql main shared bootstrap slideshow))
     end
 
     describe 'js files' do
       context "when the component's dependencies.rb does not contain .disable_auto_import" do
         it 'should list all JS files' do
           main = Volt::AssetFiles.new('/app', 'main', @component_paths)
-          expect(main.javascript(volt_app).reject{|v| v[0] != :src }).to eq([
-            [:src, "/app/volt/assets/js/jquery-2.0.3.js"],
-            [:src, "/app/volt/assets/js/volt_js_polyfills.js"],
-            [:src, "/app/volt/assets/js/volt_watch.js"],
-            [:src, "/app/bootstrap/assets/js/bootstrap.js"],
-            [:src, "/app/shared/assets/js/test2.js"],
-            [:src, "/app/slideshow/assets/js/test3.js"],
-            [:src, "/app/main/assets/js/test1.js"],
-            [:src, "/app/volt/volt/app.js"],
-            [:src, "/app/components/main.js"]
-          ])
+          expect(main.javascript(volt_app).reject{|v| v[0] != :src }).to eq(
+            [
+              [:src, "/app/volt/assets/js/jquery-2.0.3.js"],
+              [:src, "/app/volt/assets/js/volt_js_polyfills.js"],
+              [:src, "/app/volt/assets/js/volt_watch.js"],
+              [:src, "/app/bootstrap/assets/js/bootstrap.js"],
+              [:src, "/app/shared/assets/js/test2.js"],
+              [:src, "/app/slideshow/assets/js/test3.js"],
+              [:src, "/app/main/assets/js/test1.js"],
+              [:src, "/app/volt/volt/app.js"],
+              [:src, "/app/components/main.js"]
+            ]
+          )
         end
       end
 
       context "when the component's dependencies.rb contains .disable_auto_import" do
         it 'should list only the files included via the css_file helpers' do
           disabled_auto = Volt::AssetFiles.new('/app', 'disable_auto', @component_paths)
-          expect(disabled_auto.javascript(volt_app).reject{|v| v[0] != :src }).to eq([
-            [:src, "/app/volt/assets/js/jquery-2.0.3.js"],
-            [:src, "/app/volt/assets/js/volt_js_polyfills.js"],
-            [:src, "/app/volt/assets/js/volt_watch.js"],
-            [:src, "/app/disable_auto/assets/js/test1.js"],
-            [:src, "/app/volt/volt/app.js"],
-            [:src, "/app/components/main.js"]
-          ])
+          expect(disabled_auto.javascript(volt_app).reject{|v| v[0] != :src }).to eq(
+            [
+              [:src, "/app/volt/assets/js/jquery-2.0.3.js"],
+              [:src, "/app/volt/assets/js/volt_js_polyfills.js"],
+              [:src, "/app/volt/assets/js/volt_watch.js"],
+              [:src, "/app/disable_auto/assets/js/test1.js"],
+              [:src, "/app/volt/volt/app.js"],
+              [:src, "/app/components/main.js"]
+            ]
+          )
         end
       end
     end
@@ -56,11 +60,13 @@ if RUBY_PLATFORM != 'opal'
         it 'should list all the asset files in that component' do
           main = Volt::AssetFiles.new('/app', 'main', @component_paths)
 
-          expect(main.css).to eq([
-            "/app/volt/assets/css/notices.css",
-            "/app/bootstrap/assets/css/01-bootstrap.css",
-            "/app/main/assets/css/test3.css"
-          ])
+          expect(main.css).to eq(
+            [
+              "/app/volt/assets/css/notices.css",
+              "/app/bootstrap/assets/css/01-bootstrap.css",
+              "/app/main/assets/css/test3.css"
+            ]
+          )
         end
       end
 
@@ -68,10 +74,12 @@ if RUBY_PLATFORM != 'opal'
         it 'should list only the files included via the css_file helpers' do
           disabled_auto = Volt::AssetFiles.new('/app', 'disable_auto', @component_paths)
 
-          expect(disabled_auto.css).to eq([
-            "/app/volt/assets/css/notices.css",
-            "/app/disable_auto/assets/css/test1.css"
-          ])
+          expect(disabled_auto.css).to eq(
+            [
+              "/app/volt/assets/css/notices.css",
+              "/app/disable_auto/assets/css/test1.css"
+            ]
+          )
         end
       end
     end